[Attributor] Keep track of reached returns in AAPointerInfo
Instead of visiting call sites in Attribute::chenjackForAllUses, we now
keep track of returns in AAPointerInfo and use the call site return
information as required. This way, the user of
AAPointerInfo(CallSite)Argument can determine if the call return should
be visited and we do not collect them as "may accesses" in the
AAPointerInfo(CallSite)Argument itself in case a return user is found.