[Instrumentor] Fix support for globals and compile-time safe accesses
The call instrumentation was not working properly for globals and accesses
that were marked as compile-time safe (e.g., constant offset on a alloca).
This commit fixes that and removes the need for annotations.

The option OnlyReplacedPointerValues in the call instrumentation is an
optimzation to avoid passing non-pointer arguments and pointer arguments
that were not replaced by another runtime call. This is very useful for
the use case of the address sanitizer.