[Instrumentor] Fix the use of temporary allocas for runtime calls
The reuse of temporary allocas for runtime calls did not consider when
a call can use more than one alloca with the same total size. In that
case, both were trying to use the same allocas, producing invalid IR.

Now, we allow more than one alloca for each size. Each RTArgumentPack
holds the temporary allocas that are used for a particular runtime call.
Once the object is destroyed, these allocas are returned for reuse.